Break up big function
