+++ src/designer/elsa-workflows-designer/src/modules/workflow-definitions/components/models.ts
@@ -1,0 +1,16 @@
+import {Activity, ActivityDescriptor, TabDefinition} from "../../../models";
+import {ActivityInputContext} from "../../../services/activity-input-driver";
+
+export interface RenderActivityPropsContext {
+  activity: Activity;
+  activityDescriptor: ActivityDescriptor;
+  title: string;
+  inputs: Array<RenderActivityInputContext>;
+  tabs: Array<TabDefinition>;
+  selectedTabIndex: number;
+}
+
+export interface RenderActivityInputContext {
+  inputContext: ActivityInputContext
+  inputControl?: any;
+}
